+++ bundles/framework/announcements/view/index.js
@@ -2,3 +2,5 @@ export { showAnnouncementsPopup } from './AnnouncementsPopup';
 export { AnnouncementsCollapse } from './AnnouncementsCollapse';
 export { AnnouncementsContent } from './AnnouncementsContent';
 export { AnnouncementsHandler } from './AnnouncementsHandler';
+export { SingleContent } from './popup/SingleContent';
+export { PaginatedContent } from './popup/PaginatedContent';
